Modifications in response to second pull request comments
