+++ python/sqlalchemy/src/config.py
@@ -1,0 +1,8 @@
+
+
+listen_port = 8000
+db_user = 'postgres'
+db_password = 'pwd'
+database = 'yugabyte_alchemy'
+db_host = 'localhost'
+db_port = 5433
